fix fileIdをulong型で扱う

+++ Assets/HistoryViewer/Scripts/Editors/GameObjectHistoryWindow.cs
@@ -38,7 +38,7 @@ namespace Negi0109.HistoryViewer.Editors
                     foreach (var commit in _sceneGit.commits)
                     {
                         EditorGUILayout.LabelField(commit.name);
-                        if (commit.unityYaml.TryGetGameObject((int)GlobalObjectId.GetGlobalObjectIdSlow(_target).targetObjectId, out var gameObjectYaml))
+                        if (commit.unityYaml.TryGetGameObject(GlobalObjectId.GetGlobalObjectIdSlow(_target).targetObjectId, out var gameObjectYaml))
                         {
                             foreach (var componentId in gameObjectYaml.GameObject.componentIds)
                             {
